Fixed bug in pdb().
Fixed a bug in the pdb() function that occurred when the program was run
with the default OS. When the default OS was used, the pdb loop entered
in the PRE_RUNNING_STATE, without having loaded an OS. Then the program
was loaded using vm->load(...) followed by initializing the vm with
vm->init(). However, since no OS had been burned, the burn() was called
with the default OS. In burn, the memory is cleared, effectively
removing the program that had been loaded from memory.