pvm: Added cpu abstraction and requirement for examine()
The cpu abstraction steps from the new requirement for vms to implement an
examine() function. The purpose of examine() is to allow an observer of
a vm, i.e., the program using a pep vm, to unobtrusively examine its cpu
state and its memory. Unobtrusive here means that examine() can have no
side-effects on the state of the vm.